@@ -1,51 +1,56 @@
 {
-    "name": "lorebrary-api",
-    "version": "1.0.0",
-    "description": "Node.js Server for Lorebrary web",
-    "main": "index.js",
-    "scripts": {
-        "start": "node .",
-        "dev": "nodemon .",
-        "test": "jest",
-        "lint": "eslint .",
-        "pretty": "prettier --write .",
-        "prepare": "husky install"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/Two-Boys-and-a-Dream/lorebrary-api.git"
-    },
-    "keywords": [
-        "Node",
-        "JavaScript",
-        "Express"
-    ],
-    "author": "Raymond Cox",
-    "license": "MIT",
-    "bugs": {
-        "url": "https://github.com/Two-Boys-and-a-Dream/lorebrary-api/issues"
-    },
-    "homepage": "https://github.com/Two-Boys-and-a-Dream/lorebrary-api#readme",
-    "dependencies": {
-        "cors": "^2.8.5",
-        "dotenv": "^16.1.4",
-        "express": "^4.18.2",
-        "mongoose": "^7.2.2"
-    },
-    "devDependencies": {
-        "@types/jest": "^29.5.2",
-        "eslint": "^8.42.0",
-        "eslint-config-prettier": "^8.8.0",
-        "eslint-plugin-jest": "^27.2.1",
-        "husky": "^8.0.3",
-        "jest": "^29.5.0",
-        "jest-sonar": "^0.2.16",
-        "lint-staged": "^13.2.2",
-        "nodemon": "^2.0.22",
-        "prettier": "^2.8.8"
-    },
-    "lint-staged": {
-        "*.js": "eslint .",
-        "*.{js,css,md,json}": "prettier --write"
-    }
+  "name": "lorebrary-api",
+  "version": "1.0.0",
+  "description": "Node.js Server for Lorebrary web",
+  "type": "module",
+  "main": "index.js",
+  "scripts": {
+    "start": "node .",
+    "dev": "node .",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "lint": "eslint .",
+    "pretty": "prettier --write .",
+    "pretty:check": "prettier --check .",
+    "prepare": "npx simple-git-hooks",
+    "verify": "npm run pretty:check && npm run lint && npm run test"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Two-Boys-and-a-Dream/lorebrary-api.git"
+  },
+  "keywords": [
+    "Node",
+    "JavaScript",
+    "Express"
+  ],
+  "author": "Raymond Cox",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Two-Boys-and-a-Dream/lorebrary-api/issues"
+  },
+  "homepage": "https://github.com/Two-Boys-and-a-Dream/lorebrary-api#readme",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "dotenv": "^17.2.3",
+    "express": "^5.1.0",
+    "mongoose": "^7.2.2"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.39.1",
+    "@types/jest": "^30.0.0",
+    "eslint": "^9.39.1",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-jest": "^29.1.0",
+    "eslint-plugin-n": "^17.23.1",
+    "eslint-plugin-promise": "^7.2.1",
+    "eslint-plugin-security": "^3.0.1",
+    "globals": "^16.5.0",
+    "jest": "^30.2.0",
+    "prettier": "^3.6.2",
+    "simple-git-hooks": "^2.13.1"
+  },
+  "simple-git-hooks": {
+    "pre-commit": "npm run pretty",
+    "pre-push": "npm run lint"
+  }
 }
