@@ -1,10 +1,11 @@
-require('dotenv').config()
-const express = require('express')
+import 'dotenv/config'
+import express from 'express'
+import cors from 'cors'
+import mongoose from 'mongoose'
+import loreRoute from './Routes/LoreRoute.js'
+
 const app = express()
-const cors = require('cors')
-const mongoose = require('mongoose')
 const { PORT, MONGO_URL } = process.env
-const loreRoute = require('./Routes/LoreRoute')
 
 // Configs
 app.disable('x-powered-by')
@@ -14,7 +15,7 @@ app.use(cors())
 
 // Routes
 app.get('/', (req, res) => {
-    res.send('Hello World!')
+  res.send('Hello World!')
 })
 
 app.use('/lore', loreRoute)
@@ -22,12 +23,12 @@ app.use('/lore', loreRoute)
 app.listen(PORT, main)
 
 async function main() {
-    try {
-        mongoose.set('strictQuery', true)
-        await mongoose.connect(MONGO_URL)
-        console.log('Successfully connected to DB')
-        console.log(`Listening on port ${PORT}`)
-    } catch (error) {
-        console.log('error connecting to DB', error)
-    }
+  try {
+    mongoose.set('strictQuery', true)
+    await mongoose.connect(MONGO_URL)
+    console.log('Successfully connected to DB')
+    console.log(`Listening on port ${PORT}`)
+  } catch (error) {
+    console.log('error connecting to DB', error)
+  }
 }
